backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0

# ── Redis (cache + rate limiting) ──────────────────────────────────────────
redis[hiredis]>=5.0.1

# ── Error tracking (optional — only needed if SENTRY_DSN is set) ───────────
sentry-sdk[fastapi]>=2.0.0

# ── Testing ────────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
coverage==7.5.0

# ── Lint (CI only — not needed in production image) ───────────────────────
ruff==0.4.4

# ── Load testing (run locally or in CI — not in production image) ─────────
locust>=2.28.0
